# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Real-time telemetry ingestion, health scoring, and diagnostic REST API.",
        "title": "Locomotive Digital Twin API",
        "contact": {
            "name": "HackNU Team"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/alerts/{alertId}/acknowledge": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "alerts"
                ],
                "summary": "Acknowledge an alert by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Alert ID",
                        "name": "alertId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Response-api_AlertDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            }
        },
        "/api/v1/locomotives": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "locomotives"
                ],
                "summary": "List all locomotives",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.PagedResponse-api_LocomotiveDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            }
        },
        "/api/v1/locomotives/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "locomotives"
                ],
                "summary": "Get locomotive by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Locomotive ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Response-api_LocomotiveDTO"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            }
        },
        "/api/v1/locomotives/{id}/alerts": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "alerts"
                ],
                "summary": "List active (unresolved) alerts for a locomotive",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Locomotive ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.PagedResponse-api_AlertDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            }
        },
        "/api/v1/locomotives/{id}/alerts/history": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "alerts"
                ],
                "summary": "List alerts in a time range",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Locomotive ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "RFC3339 start time",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "RFC3339 end time",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.PagedResponse-api_AlertDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            }
        },
        "/api/v1/locomotives/{id}/health": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Get latest health snapshot for a locomotive",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Locomotive ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Response-sqlc_HealthSnapshot"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            }
        },
        "/api/v1/locomotives/{id}/health/history": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "List health snapshots in a time range",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Locomotive ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "RFC3339 start time",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "RFC3339 end time",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Max records (default 100)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.PagedResponse-sqlc_HealthSnapshot"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            }
        },
        "/api/v1/metrics/definitions": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "metrics"
                ],
                "summary": "List all metric definitions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.PagedResponse-api_MetricDefinitionDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "metrics"
                ],
                "summary": "Create or update a metric definition",
                "parameters": [
                    {
                        "description": "Metric definition",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.MetricDefinitionDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Response-api_MetricDefinitionDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.AlertDTO": {
            "type": "object",
            "properties": {
                "acknowledged": {
                    "type": "boolean"
                },
                "code": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "locomotive_id": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "metric_name": {
                    "type": "string"
                },
                "metric_value": {
                    "type": "number"
                },
                "recommendation": {
                    "type": "string"
                },
                "resolved_at": {
                    "type": "string"
                },
                "severity": {
                    "type": "string"
                },
                "threshold": {
                    "type": "number"
                },
                "triggered_at": {
                    "type": "string"
                }
            }
        },
        "api.LocomotiveDTO": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "display_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "last_seen_at": {
                    "type": "string"
                },
                "loco_type": {
                    "type": "string"
                },
                "registered_at": {
                    "type": "string"
                }
            }
        },
        "api.MetricDefinitionDTO": {
            "type": "object",
            "properties": {
                "crit_above": {
                    "type": "number"
                },
                "crit_below": {
                    "type": "number"
                },
                "description": {
                    "type": "string"
                },
                "display": {
                    "type": "string"
                },
                "health_weight": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "unit": {
                    "type": "string"
                },
                "warn_above": {
                    "type": "number"
                },
                "warn_below": {
                    "type": "number"
                }
            }
        },
        "api.PagedResponse-api_AlertDTO": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.AlertDTO"
                    }
                },
                "success": {
                    "type": "boolean"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "api.PagedResponse-api_LocomotiveDTO": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.LocomotiveDTO"
                    }
                },
                "success": {
                    "type": "boolean"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "api.PagedResponse-api_MetricDefinitionDTO": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.MetricDefinitionDTO"
                    }
                },
                "success": {
                    "type": "boolean"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "api.PagedResponse-sqlc_HealthSnapshot": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/sqlc.HealthSnapshot"
                    }
                },
                "success": {
                    "type": "boolean"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "api.Response-any": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "api.Response-api_AlertDTO": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/api.AlertDTO"
                },
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "api.Response-api_LocomotiveDTO": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/api.LocomotiveDTO"
                },
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "api.Response-api_MetricDefinitionDTO": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/api.MetricDefinitionDTO"
                },
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "api.Response-sqlc_HealthSnapshot": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/sqlc.HealthSnapshot"
                },
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "sqlc.HealthSnapshot": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "factors": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "locomotive_id": {
                    "type": "string"
                },
                "metrics_snap": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "score": {
                    "type": "integer"
                },
                "ts": {
                    "type": "string"
                }
            }
        }
    }
}